+++ hubdle-settings-gradle-plugin/main/kotlin/com/javiersc/hubdle/settings/HubdleSettingsExtension.kt
@@ -5,7 +5,9 @@ import com.javiersc.hubdle.settings.extensions.HubdleBuildScanExtension
 import javax.inject.Inject
 import org.gradle.api.Action
 import org.gradle.api.model.ObjectFactory
+import org.gradle.api.provider.Property
 import org.gradle.kotlin.dsl.newInstance
+import org.gradle.kotlin.dsl.property
 
 @HubdleSettingsDslMarker
 public open class HubdleSettingsExtension @Inject constructor(objects: ObjectFactory) {
@@ -23,4 +25,6 @@ public open class HubdleSettingsExtension @Inject constructor(objects: ObjectFac
     public fun buildScan(action: Action<HubdleBuildScanExtension> = Action {}) {
         action.execute(buildScan)
     }
+
+    public val useOnAllProjects: Property<Boolean> = objects.property<Boolean>().convention(true)
 }